LayoutTests/fast/dom/insertedIntoDocument-iframe-expected.txt

CONSOLE MESSAGE: line 1: TypeError: Argument 1 ('child') to Node.removeChild must be an instance of Node
PASS
